Move handle cj final to WalletCoinJoinClient
