+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/CoinJoinManager.cs
@@ -321,162 +321,6 @@ public class CoinJoinManager : BackgroundService
 		return coinJoinClientStates.ToImmutable();
 	}
 
-	private async Task HandleCoinJoinFinalizationAsync(CoinJoinTracker finishedCoinJoin, ConcurrentDictionary<WalletId, CoinJoinTracker> trackedCoinJoins, ConcurrentDictionary<IWallet, TrackedAutoStart> trackedAutoStarts, CancellationToken cancellationToken)
-	{
-		var wallet = finishedCoinJoin.Wallet;
-		var batchedPayments = wallet.BatchedPayments;
-		CoinJoinClientException? cjClientException = null;
-		try
-		{
-			var result = await finishedCoinJoin.CoinJoinTask.ConfigureAwait(false);
-			if (result is SuccessfulCoinJoinResult successfulCoinjoin)
-			{
-				CoinRefrigerator.Freeze(successfulCoinjoin.Coins);
-				batchedPayments.MovePaymentsToFinished(successfulCoinjoin.UnsignedCoinJoin.GetHash());
-				await MarkDestinationsUsedAsync(successfulCoinjoin.OutputScripts).ConfigureAwait(false);
-				wallet.LogInfo($"{nameof(CoinJoinClient)} finished. Coinjoin transaction was broadcast.");
-			}
-			else
-			{
-				wallet.LogInfo($"{nameof(CoinJoinClient)} finished. Coinjoin transaction was not broadcast.");
-			}
-		}
-		catch (UnknownRoundEndingException ex)
-		{
-			// Assuming that the round might be broadcast but our client was not able to get the ending status.
-			CoinRefrigerator.Freeze(ex.Coins);
-			await MarkDestinationsUsedAsync(ex.OutputScripts).ConfigureAwait(false);
-			Logger.LogDebug(ex);
-		}
-		catch (CoinJoinClientException clientException)
-		{
-			cjClientException = clientException;
-			Logger.LogDebug(clientException);
-		}
-		catch (InvalidOperationException ioe)
-		{
-			Logger.LogWarning(ioe);
-		}
-		catch (OperationCanceledException)
-		{
-			if (finishedCoinJoin.IsStopped)
-			{
-				wallet.LogInfo($"{nameof(CoinJoinClient)} stopped.");
-			}
-			else
-			{
-				wallet.LogInfo($"{nameof(CoinJoinClient)} was cancelled.");
-			}
-		}
-		catch (UnexpectedRoundPhaseException e)
-		{
-			// `UnexpectedRoundPhaseException` indicates an error in the protocol however,
-			// temporarily we are shortening the circuit by aborting the rounds if
-			// there are Alices that didn't confirm.
-			// The fix is already done but the clients have to upgrade.
-			wallet.LogInfo($"{nameof(CoinJoinClient)} failed with exception: '{e}'");
-		}
-		catch (WabiSabiProtocolException wpe) when (wpe.ErrorCode == WabiSabiProtocolErrorCode.WrongPhase)
-		{
-			// This can happen when the coordinator aborts the round in Signing phase because of detected double spend.
-			wallet.LogInfo($"{nameof(CoinJoinClient)} failed with: '{wpe.Message}'");
-		}
-		catch (Exception e)
-		{
-			wallet.LogError($"{nameof(CoinJoinClient)} failed with exception: '{e}'");
-		}
-		finally
-		{
-			batchedPayments.MovePaymentsToPending();
-		}
-
-		// If any coins were marked for banning, store them to file
-		if (finishedCoinJoin.BannedCoins.Count != 0)
-		{
-			foreach (var info in finishedCoinJoin.BannedCoins)
-			{
-				CoinPrison.Ban(info.Coin, info.BanUntilUtc);
-			}
-		}
-
-		NotifyCoinJoinCompletion(finishedCoinJoin);
-
-		// When to stop mixing:
-		// - If stop was requested by user.
-		// - If cancellation was requested.
-		if (finishedCoinJoin.IsStopped
-			|| cancellationToken.IsCancellationRequested)
-		{
-			NotifyWalletStoppedCoinJoin(wallet);
-		}
-		else if (await wallet.IsWalletPrivateAsync().ConfigureAwait(false))
-		{
-			NotifyCoinJoinStartError(wallet, CoinjoinError.AllCoinsPrivate);
-			if (!finishedCoinJoin.StopWhenAllMixed)
-			{
-				// In auto CJ mode we never stop trying.
-				ScheduleRestartAutomatically(wallet, trackedAutoStarts, finishedCoinJoin.StopWhenAllMixed, finishedCoinJoin.OverridePlebStop, finishedCoinJoin.OutputWallet, cancellationToken);
-			}
-			else
-			{
-				// We finished with CJ permanently.
-				NotifyWalletStoppedCoinJoin(wallet);
-			}
-		}
-		else if (cjClientException is not null)
-		{
-			// - If there was a CjClient exception, for example PlebStop or no coins to mix,
-			// Keep trying, so CJ starts automatically when the wallet becomes mixable again.
-			ScheduleRestartAutomatically(wallet, trackedAutoStarts, finishedCoinJoin.StopWhenAllMixed, finishedCoinJoin.OverridePlebStop, finishedCoinJoin.OutputWallet, cancellationToken);
-			NotifyCoinJoinStartError(wallet, cjClientException.CoinjoinError);
-		}
-		else
-		{
-			wallet.LogInfo($"{nameof(CoinJoinClient)} restart automatically.");
-
-			ScheduleRestartAutomatically(wallet, trackedAutoStarts, finishedCoinJoin.StopWhenAllMixed, finishedCoinJoin.OverridePlebStop, finishedCoinJoin.OutputWallet, cancellationToken);
-		}
-
-		if (!trackedCoinJoins.TryRemove(wallet.WalletId, out _))
-		{
-			wallet.LogWarning("Was not removed from tracked wallet list. Will retry in a few seconds.");
-		}
-		else
-		{
-			finishedCoinJoin.WalletCoinJoinProgressChanged -= CoinJoinTracker_WalletCoinJoinProgressChanged;
-			finishedCoinJoin.Dispose();
-		}
-	}
-
-	/// <summary>
-	/// Mark all the outputs we had in any of our wallets used.
-	/// </summary>
-	private async Task MarkDestinationsUsedAsync(ImmutableList<Script> outputs)
-	{
-		var scripts = outputs.ToHashSet();
-		var wallets = await WalletProvider.GetWalletsAsync().ConfigureAwait(false);
-		foreach (var k in wallets)
-		{
-			var kc = k.KeyChain;
-			var state = KeyState.Used;
-
-			// Watch only wallets have no key chains.
-			if (kc is null && k is Wallet w)
-			{
-				foreach (var hdPubKey in w.KeyManager.GetKeys(key => scripts.Any(key.ContainsScript)))
-				{
-					w.KeyManager.SetKeyState(state, hdPubKey);
-				}
-
-				w.KeyManager.ToFile();
-			}
-			else
-			{
-				k.KeyChain?.TrySetScriptStates(state, scripts);
-			}
-		}
-	}
-
 	private void NotifyWalletStartedCoinJoin(IWallet openedWallet) =>
 		StatusChanged.SafeInvoke(this, new WalletStartedCoinJoinEventArgs(openedWallet));
